fix(core): Only log relevant details of description (#4456) (#4457)
When we encounter a description that should have application restrictions
but doesn't, we log a warning including the full description JSON. In some
cases (ex: Kubernetes) this description may contain sensitive information
that ideally wouldn't be logged.

Ideally really sensitive values would come from artifacts so they aren't
in the pipeline JSON at all, but let's nonetheless cut back on how much
we are logging here to reduce the chance of sensitive information ending
up in the logs.

In particular, it's not clear how much the full description would help in
debugging, as generally what is important is the class of the description
(to check whether it implements ApplicationNameable or ResourcesNameable)
and and the account, both of which will still be logged after this change.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>