chore(mergify): Update mergify comment
The comment was incorrect; depending on the repo, there are a few
branches that use both Travis and Github actions. Also, this comment
matches what I put in all the other repos, so this makes the
mergify config here identical to the other ones, which makes things
easier for bulk changes.