refactor(core): Move some things back to halyard config
In a recent PR I moved the server.port and server.address settings
into the base config to reduce what's in the Halyard config.

After further reflection, we don't want these to be overridable
easily in kleat anyway as I don't think there's a common use case
for this.  (And in any case, it's easier to override this by just
overriding these in the clouddriver-local.yml than to need to go
through the services.* settings in spinnaker.yml.)

So let's avoid bringing these settings along for the ride, and
just leave them in the halconfig where they will still help
halyard users. This restores this block of clouddriver.yml to
what it was before the recent change.