fix(config): Remove default from redis-ro
Change redis-ro to redisRo to match the casing in the rest of
spinnaker.yml.  (I don't want to take on the project of updating that
file to kebab case right now.)

Also remove the default; we're going to require that redisRo be defined
in spinnaker.yml if you're running HA (even if it just points to the
regular redis). Maybe later we'll deafult it, but it's safer to add
a default later than to remove one later.