feat(kubernetes): Add support for kubectl create
Some manifest need to be applied with kubectl create; one example
is manifests that specify a generateName instead of a name. In this
case, we can't apply or replace, as both of these operations require
looking up the existing manifest which is impossible when the name
changes each time.

This commit adds support to KubernetesV2Credentials to create a
manifest using kubectl create operation.