feat(kubernetes): Apply manifests with generateName using create
Manifests that have a generateName instead of a name need to be
created using a create operation, as other operations require
looking up the existing manifest to reconcile changes, but this
is not possible if we don't have the manifest's name.

This change updates CanDeploy to check if the manifest has a
generateName (and no name, as this would take priority), and to
always use 'kubectl create' if that is the case.