test(kubernetes): Add tests to metric recording on kubectl calls
The runAndRecordMetrics funcgtion in KubernetesV2Credentials is
very complex and has no tests. Before simplifying it, add some
tests to document and validate the current behavior.

This commit also updates the constructor of KubernetesManifest to
set the metadata to an empty map. In real code, we always create a
KubernetesManifest by deserializing, so the metadata is set to a
map of whatever was on the incoming manifest. But it's often useful
in tests to be able to create a manifest and set a few fields on it;
before this change, this wasn't possible because metadata was always
null. There were functions to set various keys on the metadata, but
these would all fail due to the null metadata.