fix(kubernetes): Fix some issues with kubectl metrics
Cleaning up the code in the last commit surfaced a few issues
that this commit cleans up.

(1) We should add a failure tag if a KubectlException is thrown
(2) Metric tags are intended to be short descriptors, so we shouldn't
add the text of an error message. Remove that message from the cause
tag and just include the class of the exception.
(3) There's no reason to separately tag 'kind' and 'kinds' depending if
there are more than 1 or not; just always add 'kinds' with the full list
of kinds. Also sort it for consistency in the metric tags.

There's probably a bit we can do to futher clean up the exception handling
(deciding when to wrap the exception, etc.) and deduplicate the two catch
blocks but leaving that s is for now.