refactor(kubernetes): Improve error abstraction in kubectl
When parsing the output of kubectl as JSON, a few different kinds of
errors can happen:
(1) An IOException if we encounter an error reading from the output
stream
(2) A JSONSyntaxException if the call to gson.fromJson fails
(3) An IllegalStateException if the JSON format is unexpected (ex:
if we get a list back instead of an object and call .beginObject()
on it).

Currently we're propagating all these exceptions as is. Let's continue
letting the IOException propagate, as this can be handled by the job
executor in the same way it handles any other error reading output.

Let's wrap the JSONSyntaxException and IllegalStateException in a
KubectlException to provide relevant context. This is how we handle
a JSONSyntaxException other places in this file when we parse
non-streaming responses. (Potentially at some point this could be
a specific subclass of KubectlException but for now just being
consistent with what happens elsewhere in the file.)

This means that now the error surface out of the KubectlJobExecutor
for a JSON parse exception will be the same regardless of whether
this happened in the streaming or non-streaming parser.