refactor(jobs): Throw exceptions appropriate to the abstraction
The JobExecutor class throws RuntimeException in a few places; let's
replace this with throwing a JobExecutionException with information about
the job that failed.

Let's also take the opportunity to document the contract about when the
JobExecutor throws an exception vs. when it returns a JobResult with a
status of FAILED.