refactor(kubernetes): Remove NoResourceTypeException
This exception was added in PR2344, with the goal of in some cases
catching exceptions where a type was not recognized by the API
(for example if a built-in kind is not present on an old Kubernetes
cluster). It appears, however, that the change in that PR never
actually worked, as all relevant calls to kubectl go through
runAndRecordMetrics, which wraps any exception that is not a
KubectlException in a KubectlException. So the catch blocks
added in that PR will never be triggered as we'll always see
a KubectlException at that point.

Probably what should have been done is to make NoResourceTypeException
extend KubectlException (or potentially to have it special-cased in
runAndRecordMetrics). In the intervening time since that PR, we've
added a check the first time we try to read a kind to make sure that
it's readable (mainly for permissions issues, but it would also
address this case), so I'm not sure there's much value in actually
fixing this. (In any case, the behavior hasn't worked in the two years
since it was added and we have not heard reports of people running into
unrecognized kinds, except briefly when the readable check noted earlier
was broken.)

This commit then does the following:
(1) Removes all the catch blocks for NoResourceTypeException that are
unreachable
(2) Removes NoResourceTypeException itself; given that calling code is
not distinguishing this particular exception, there's not much benefit
to throwing it vs. a generic KubectlException (particularly given that
we're relying on string parsing to do so, and that we're only throwing
it in some of the methods, not others).
(3) It makes the constructor of KubectlException package-private instead
of protected.