refactor(kubernetes): Fix a few places where we catch KubectlException
In the metrics caching agent, we catch a KubectlException and dig into
the message to see if it's because metrics aren't available, then return
an empty collection in that case. Let's just have topPod directly handle
that so we don't leak the implementation detail of using kubectl (and
the specific message); this makes sense for the other two places we call
this function as well.

In KubernetesV2CachingAgent, we catch, log, then re-throw an exception.
This is a vestige of when we used to just return an empty list after
logging, but that was change to propagate the exception, so there's
not really much additional benefit from also logging.