refactor(kubernetes): Don't wrap kubectl exceptions
Let's no longer wrap all exceptions that pass through runAndRecordMetrics
in a KubectlException. While it does have the benefit of providing
some context at that level of abstraction, it is confusing that we
sometimes directly throw that in the KubectlJobExecutor and sometimes
create one by wrapping it here.

I think at some point we may want to create a new KubernetesAPIException
here that wraps anything below (either a KubectlException of a client
library exception), but for now let's just propagate whatever was
thrown below.

The one thing to be careful about is that this change doesn't break
places where we're catching a KubectlException. I've looked through
and in all cases the code is catching to deal with some expected
(but non-successful) kubectl call (ex: a missing resource, logs
not there, etc.). All of these remain KubectlExceptions (and in
some cases we should potentially even make more specific ones and
catch that specific one). What we'll no longer be catching is things
like a JobExecutionException where we fail to run the job, etc, which
I think actually is better in this case.