refactor(kubernetes): Operations should return the created manifest
When creating a manifest, the result is returned by the API, but
is currently ignored in most cases by clouddriver. It is useful
to use the returned manifest because there are times where it has
updated information. One particular example is when we don't specify
a namespace on our manifest; in that case, the manifest will be
applied to the default namespace that is configured.

Currently we solve this problem by separately looking up the default
namespace before applying the manifest, but it would be easier to just
apply it and then check what kubernetes default it to. This change
will allow us to do that in a future commit.