refactor(kubernetes): Don't look up default namespace
There are two places where we look up the default namespace
unnecessarily. The first is on every call to kubectl so that we
can pass in a "--namespace" flag; if we just leave that flag out,
we'll get the default namespace already. The second is when
deploying a manifest without a namespace; in that case, we look up
the default namespace so that we can apply it to the manifest before
deploying. Similarly, if we just deploy the manifest as is, it will
get the default namespace (which we'll see on the result from the
API).

I tested that this works both for the serviceAccount case and for
the regular kubeconfig case (testing the serviceAccount by ssh-ing
to a pod set up this way to check that kubectl already detected the
default).

This deletes a lot of unecessary code. When we move to use the client
library, we may need to solve this again as I'm not sure that it will
always detect a default namespace, but this code was complicated enough
and coupled pretty tightly to the kubectl implementation that I think
it will be better to start from scratch at that point.  (This was also
broken, as the command to get the default namespace from the kubeconfig
appears to fail every time so you always get 'default' regardless of what
is set.)