feat(kubernetes): Replace deploy strategy acts as upsert
Currently if using the strategy.spinnaker.io/replace annotation,
the manifest will be applied with kubectl replace. This will fail
the first time the pipeline is run as the object won't exist and
kubectl replace fails if there is no existing resource to replace.
(There is an open issue in kubernetes to make replace act as an
upsert, but as of now it just as workarounds to try the replace
and then create if it fails.)

As this is a common issue, Spinnaker should be able to make this
easier. This commit updates the behavior on doing a replace to
fall back to creating the resource if the replace call fails due
to the resource not existing.