refactor(kubernetes): A few minor improvements to KubernetesKind
Use the new NonnullByDefault annotation, and remove a bunch of
now redundant Nonnull annotations.

Also, simplify fromString to use a Guava splitter, which avoids
needing to use an array, and can simplify the logic. This is the
only part of this PR that isn't a simple rewrite of one function
to another, but the modified function here has a number of good
tests on it.