refactor(kubernetes): Replace Collections.empty... with Guava
In a lot of places we use Collections.emptyList(), emptyMap(),
While it's not in the name of the function, the collection that
is returned is actually immutable.

Now that we've started adding lots of uses of Guava's immutable
collections in the Kubernetes provider, it makes sense for these
to also use the Guava types. This has a few advantages:
* It reduces the burden on new contributors to figure out which
of the two ways of doing the same thing they should use.
* The more we have Immutable collections floating around, the
more likely people will find that code in examples and use them
in their contributions.
* It's much more clear that the returned type is immutable
* In some cases, this allows us to also upgrade the return type
of the entire function to an immutable collection (when the only
return path was the one we just updated)

This change is safe because, as noted above, the collections were
already immutable. (Wheras changing a mutable collection to immutable
requires some care that it's not being mutated somewhere downstream.)