refactor(kubernetes): Replace singleton collections with Guava
The rationale is in the last commit; this commit does the same
thing the last commit did with empty collections and does it for
singleton collections. Again, as these are already immutable, the
change is safe and actually makes it easier in some cases for the
caller to realize the returned collection is immutable.