refactor(google): Convert GoogleCommonSafeRetry to java
In addition to converting this to Java, this commit also allows
updates the bare types to be generic, so that the methods return
something more strongly-typed than an object, which is very useful
for strongly-typing any downstream code without casts.

It also updates the class to no longer be a superclass of the
places that use it; it instead encapsulates its logic and
consumers can create one and store it as an instance variable.

As part of this change, instead of having consumers inject in
an exception factory, this class just throws a GoogleAPIException,
which consumers can then wrap as they see fit.

(I came across this while fixing a bug in front50 that uses this,
unfortunately by importing clouddriver as a library, and was unable
to do so in a typesafe way without this change.)

This class is pretty robustly tested, as the GCE and appengine
operations all run through this class and often assert that the
right type of exception is thrown (in addition to the tests
directly on this class). For example, when I had a bug in the
ResourceInUseException code, it was caught by a GCE test.

There are two changes to the tests on this class that are worth
noting:
* In one case, the exception type is changed from IllegalStateException
to GoogleAPIException. This was just because the tests was using
IllegalStateException in its factory.
* In another case we've changed from returning GoogleJsonResponseException
to a GoogleAPIException. I'm fairly certain this was originally a bug
as we were incorrectly throwing that exception instead of storing it
and processing it through the logic to decide what to return. From
reading the places this function is called, it doesn't look like we are
catching a GoogleJsonResponseException anywhere so I don't think this should
affect any consumers.

Finally, the isRetryable method has not been used in ~2 years; at some point
a refactor removed the call to it (and its logic to retry all 5xx errors). As
I don't want to re-introduce that now without a reason, I just removed the
function and its tests.