refactor(appengine): Remove success field from appengine
Ideally the retry logic would not accept a list of statuses
to consider success, as this means that we need it to be
nullable because if we get one of those statuses we need to return
but don't have anything to return. (This might be more resonable
if we had separate calls for operations that returned a value and
operations that didn't; in that case the ones that don't return
a value could more reasonably accept this parameter.)

As this isn't actually used anywhere in appengine, let's just delete
it to clean up the code a bit.