refactor(google): Clean up handling of error details
The fromGoogleJsonException method was doing two things:
(1) parsing out the details on the exception to figure out what
we know about it
(2) deciding what type of GoogleApiException to throw

This was alredy a bit complicated but I'm about to add another
subclass of GoogleApiException so things are going to get out of
hand if we don't fix it. Create and ErrorDetails private class
that figures out what we know about an exception and puts it in
a standard format, then let fromGoogleJsonException base its
decision on this class.