refactor(google): Return NotFoundException from API calls
The code in front50 that uses GoogleCommonSafeRetry was relying
on getting back the raw GoogleJSONResposeException, but only
because it was digging into the exception to figure out if it
was a 404 to respond differently. Let's instead encapsulate this
and return a GoogleApiException.NotFoundException in this case so
we can catch that.