refactor(google): Make GoogleApiException checked
This might be controversial, and definitely interested in feedback
on this one, but the expectation at this point is that all callers
are catching this and wrapping it in something that describes what
they were doing when the exception happened. Because we're expecting
them to catch it, having this be a checked exception actually makes
it easier to ensure we didn't miss a place...but I also could see
the point that there's not much you can do about this beyond wrap
it so maybe it shouldn't be checked.