refactor(artifacts): Make artifact accounts immutable and null-safe
There are a lot of config properties that can be set for artifacts.
In general, they don't care about the difference between unset and
null values, but this is enforced at various points of rather than
at the API boundary. In order to make it more clear that this is
the case, update the configuration classes to be immutable and to
translate null strings to empty strings upon creation. This removes
the ability of any consuming code to be able to see an null value
in these string properties.

In particular, we make all fields on artifact accounts immutable,
and replace the setters with a constructor. The ConstructorBinding
annotation is what tells Spring to bind properties using the
constructor (without it, it will look for setters and fail).
We also provide a Builder for convenience (primarily for testing)
to avoid needing to use the many-arg-constructor.

As the constructor will be called by Spring, which does allow null
properties we need to treat all input as potentially null and
convert to non-null in the constructor.