refactor(artifacts): Make ArtifactCredentials null-safe
Now that the artifact accounts are immutable, the artifact credentials
that read from them can be made NonNullByDefault. This commit also
removes a number of now-redundant null-checks on its fields, replacing
them with .isEmpty() on a now guaranteed to be non-null string.

This commit also updates one of the interface methods to return an
ImmutableList (which all but one implementation was effectively doing
by returning Collections.singletonList), and makes the interface
NonnullByDefault.

There was a single implementation of download() that was not null-safe.
I've updated it to eagerly throw an exception; all callers of that method
that was updated immediately pass the output to IOUtils.copy which will
throw a null pointer exception. This change will simply provide more
context around the exception.