fix(gce): Fix cluster caching on the SQL backend (#4607)
This is fixed on master as #4570, but I don't think it's safe enough to cherry-pick that. This is a significantly hackier fix, but it's safer since it only affects users using the SQL backend, which apparently isn't working anyway.

This is against the release-1.20.x branch, and then I'll cherry-pick it back to 1.19 and 1.18.

Co-authored-by: Ethan Rogers <ethanfrogers@users.noreply.github.com>