refactor(artifacts): Remove mutations of artifacts
In an upcoming kork release, artifacts will become immutable to
prevent confusing bugs where (primarily in orca) we are mutating
copies of what we think we are mutating. There are a small number
of places in clouddriver where artifacts are mutated, but these
can be replaced by returning a copy with the desired changed.