refactor(kubernetes): Avoid mutating artifact metadata
There's only one place where we mutate the artifact metadata
map, and it's in the Kubernetes provider. Rather than get an
artifact then add the metadata later, let's just move the logic
to add that metdata to the place where we create the artifact,
avoiding the need to mutate the metadata.