refactor(artifacts): Update artifact code to use new metadata accessor
The accessor to get the full metadata of an artifact as a map is deprecated;
replace calls to it with the accessor for a specific key.

The reason the function to return the entire map is deprecated is so that we
can eventually make the map immutable; it's much safer to make that change
if we've first updated callers to be unaware that the map exists at all
an instead give them a function to get the information they want.

In most cases this actually simplifies the calling code as we can remove
null-checks that are now handled by the exposed API function.