fix(tencent): Fix compilation error due to import order
This file is not compiling for my version of the JDK (openjdk 1.8.0_242)
because we are importing the inner class before we import the Data annotation
that the class is annotated with. It looks like this may have been fixed in
later versions of the JDK, but let's avoid others running into this issue
by fixing it.

Rather than switch the imports (which would be restored by spotlessApply)
let's just not import the inner class and qualify references.