refactor(core): Fix warning fetching operation version
The code to fetch the version of an operation catches any
Exception. The only actual exception that can occur is a
null-pointer exception of credentials.getOne returns null.
Let's just make the code null-safe and remove all catch and
warning logic.

One of the main motivations of this is that ProviderVersion is
no longer even relevant now that the Kubernetes V1 provider has
been removed from the code. Logging a warning related to being
unable to fetch the provider version just confuses debugging;
if the account doesn't exist, that itself is the problem, so
we should note that in the warning, rather than that we couldn't
fetch the version.