refactor(kubernetes): Clean up EnableDisableManifestDescription
Mostly the goal here is to make loadBalancers Nonnull, which
required overriding the lombok-generated accessor so that we can
translate a null that's passed in to an empty list. While here,
make the list immutable.

Also add some tests, particularly on deserialization.

I did *not* make the class immutable because one must draw some
line at which to stop refactoring.