fix(kubernetes): Fail with a useful message if manifest not found
This commit adds nullity annotations to a number of functions
related to the disable manifest stage so that we can better
understand what can be null.

It then throws a useful error message when a requested manifest
is null, preventing what was a downstream error far from the
actual issue (and that had less information about the error).