feat(plugins): move DescriptionValidator to clouddriver-api
This allows plugins to implement DescriptionValidators via clouddriver-api. Unfortunately Spring's Errors class is used in the contract of DescriptionValidator.validate() and we can't put that dependency in clouddriver-api.
This PR creates ValidationErrors which hides Errors from Spring. ValidationErrors is just an interface that defines the reject methods in Errors. DescriptionValidationErrors now implements ValidationErrors as well.
All existing validators had to get their validate method updated to use ValidationErrors instead of Errors. Some tests also needed minor changes. That is the bulk of the changes and they are all derivative changes. The core changes are in DescriptionValidator and DescriptionValidationErrors.