fix(cats/sql): Fixes CloudDriver SQL Cache from failing due to id or agent being too long to be stored.
Fixes github.com/spinnaker/spinnaker/issues/5600. The theoretical max length of an id or agent is much larger than we can store in columns that are primary or index keys. Therefore, in order to solve the issue I moved away from storing the id and agent in key columns. I've introduced new columns that will be used as keys that store hashes of the ids and agents instead. I made no changes to the contracts between the SQL Cache and those who depend on it, all the changes are internal to the SQL Cache.

I also took the liberty of adding more javadoc style comments to some of the functions in the SQL Cache. Even though some of them are simple, I still think it's beneficial to introduce it since if the functions do become more complex then whoever updates them will need to update the comment.