test(kubernetes): Add integration tests on caching views
There are currently almost no tests on any of the data providers
in the Kubernetes provider. Before doing significant work in this
area, we should add some test coverage to validate that the upcoming
changes don't break anything.

As a lot of the classes are tightly-coupled without clearly defined
interfaces between them, I decided to write an integration test that
runs the caching agent for a few resources, then calls the data
providers to ensure that the resulting data is as expected.

This tests the contract that should mostly stay consistent during
the upcoming refactor. Even if we change what data is stored in the
cache, we'll need to make sure that these providers still return
the same data after a caching cycle is run.

Eventually we should write more unit tests, once there are stronger
contracts between these providers and between them and the cache.
But at this point I think a test at this level will give the most
confidence that the upcoming changes don't break anything.

This is a very large class, as I only wanted to set up the data
once, and the various providers all return different view of the
same data so they can share some of the assert functions. Maybe
there's a better way to split this up, but for the moment this
seemed reasonable.

I used SoftAssertions so that you get a report of all the failures
for each test at the end (rather than failing at the first issue)
which should make debugging easier.

Finally, I had to make a few classes public. These aren't really
intended to be used widely, but it's impossible to appropriately
create a new KubernetesV2Credentials without them. Eventually
we could/should make it easier to create said credentials for
testing purposes, but for now this seemed better than making the
test less realistic.  (Also as these are Spring components, and
Spring can find package-private things, they are in a sense a bit
exposed already. In fact the reason we need them exposed is because
we're wiring up things as Spring would on startup.)