test(kubernetes): Address code review comments in tests
Add a test that passes details=false to getCluster, which was missing
before, and add a comment noting that the version of the call without
the flag defaults it to true.

Fix a potential exception in the load balancer assertions (so we
just assert the size is wrong and move on instead of getting a
NoSuchElementException) that I ran into when testing the above
change.

Remove a mistakenly copied comment.