test(kubernetes): Invert boolean so it matches method under test
It was confusing that the cluster assertions expected a flag summary
which is false when details are included, but the methods we're testing
expect a flag includeDetails which is true when details are included.

Update the assertion function to take a flag includeDetails (which is
the inverse of summary) for clarity.