refactor(kubernetes): Improve typing in view classes
In a number of places we store data in these classes as the interface
type rather than as the concrete type, and then need to cast to the
Kubernetes-specific implementation. Instead, let's keep everything
concrete in the Kubernetes classes to avoid this casting.

In some cases, using the interface type can be useful (ie, if we
don't want certain implementation-specific functions to be visible).
But in this case the primary use for these data classes is serializing
them and sending them to another microservice (and eventually to deck).
Regardless of the type we are storing it in, we are still fully serializing
the instance fields, so the effective contract is that callers expect
to find Kubernetes-specific fields here. (That's the reason the tests
were casting to the specific type, so that they could verify things in
the serialized results.)