refactor(kubernetes): Inline some functions in data provider
As a first pass to simplifying the Kubernetes data providers, inline
a few functions. While this makes the calling function longer, it
will be simpler because we won't be passing as many custom map
objects across function boundaries. As the later commits will show,
we'll also better be able to simplify things when the logic is all
in one function.