refactor(kubernetes): Inline a variable in cluster provider
A small commit, inlining a variable that is only used in one place
and can be just created in place rather than stored as a local
variable.

Also removed the needless null filtering in the return from the
function; the only things in this set are the results of calling
new KubernetesV2Cluster(...) which can never be null.