refactor(kubernetes): Clean up null safety in KubernetesCacheUtils
There are a lot of unnecessary null checks in KubernetesCacheUtils
and in code that calls those functions, because pretty much every
function assumes a Collection<CacheData> might contain null elements.

At first, I had this commit filtering the results from all calls
to the cache (on the assumption that the cache might return a
collection containing null elements if you requested an id that
didn't exist). But I checked all of the current implementations
of Cache, and all of them ensure filter nulls from the result.
(In the SQL case we even know this statically as it returns a
MutableCollection<CacheData> rather than a MutableCollection<CacheData?>.

To avoid the unnecessary duplicate work, just remove filtering of
the Collection<CacheData> that we get back from the cache. Then
as long we we don't ourselves create a Collection<CacheData> with
null elements, we can be sure that all such collections don't have
nulls.

This greatly simplifies the code, and also likely has performance
benefits as we no longer need to iterate over every result from
calling the cache, just to create a new collection.

Finally, this commit does not fully update all downstream logic to
remove null checks.  We can update those places as we change them.