refactor(kubernetes): Replace full cache data with keys instead
KubernetesV2ServerGroupCacheData requires a full Collection<CacheData>
of the load balancer data, but only ever uses the associated keys.
Instead, just require the actual keys, which means that callers can
avoid doing extra lookups to the cache, simplifying the logic and
improving performance.

Note that this commit does not actually change the calling logic to
stop fetching the no-longer-needed CacheData objects, it just pushes
the CacheData::getId to the caller. That simplification will come
later.

In the case of serverGroupManagerKeys, we're currently requiring a
parsed key. Encapsulate this logic better (and match what we now do
for load balancers) by accepting the key as a string and parsing the
keys as needed.