refactor(kubernetes): Fix type safety and null safety
resourceModelFromCacheData is very unsafe; it does an unchecked
cast and can also return null. The goal of that function is to
allow custom overrides for how to convert cache data into a
ServerGroup or a ServerGroupManager.  (We have not implemented
the ability to customize other types.)

The problem is that we are checking if the handler is of the
generic ModelHandler instead of the specific type we are
interested in.  We already have ServerGroupHandler and
ServerGroupManagerHandler to handle these two specific cases.
Split the generic resourceModelFromCacheData into to specific
ones to handle these two cases.  By checking if the handler
implements the specific case, we can avoid unchecked casts.

Finally, rather than return null if the handler doesn't specify
an override, just use the default implementation. The goal here
is only to allow people to override this if they want; if they
don't override, we can just use the default.